+++ src/cif/dto/create-users.dto.ts
@@ -1,4 +1,5 @@
 // src/users/dto/create-user.dto.ts
+import { Decimal } from '@prisma/client/runtime/library';
 import { IsString, IsOptional, MinLength } from 'class-validator';
 
 export class CreateUserDto {
@@ -13,5 +14,5 @@ export class CreateUserDto {
     password: string;
 
     @IsOptional()
-    balance?: bigint;
+    balance?: Decimal;
 }
